Se completaron los cambios solicitados por el cliente para iniciar producción el 10-07-2024.
- Arreglar descuento para que sean quetzales en lugar de porcentaje en creacion de venta
- Dejar cliente fijo o si da tiempo eliminar la referencia de cliente.
- Detalle de venta mostrar precio de compra de los productos
- Reemplazar precio sugerido en modulo de productos por precio de compra
- Cambiar orden de %Ganancia por precio venta (De ser posible eliminar campos de porcentaje de ganancia y precio sugerido)
- Agregar a detalle de producto el precio de compra
- Eliminar enlace a pagina web
- Modificar reporte de inventario para que se muestren precio de compra y venta unitarios y modificar titulo si da tiempo
- Arreglar Poliza de venta individual
- Saltarse el paso de cantidad en agregar producto
- Cambiar el focus del inicio al codigo del producto
- Quitar opcion de buscar cliente y campo de nit del formulario de creacion de factura

+++ Backend/mundo/src/main/java/com/aglayatech/mundo3/model/DetalleFactura.java
@@ -15,6 +15,7 @@ import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @ToString
@@ -28,9 +29,9 @@ public class DetalleFactura implements Serializable {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long idDetalle;
 	private Integer cantidad;
-	private Double subTotal;
-	private Double descuento;
-	private Double subTotalDescuento;
+	private BigDecimal subTotal;
+	private BigDecimal descuento;
+	private BigDecimal subTotalDescuento;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "id_producto")
